refactor: Divide layer
