+++ src/main/java/view/InputView.java
@@ -1,0 +1,4 @@
+package view;
+
+public class InputView {
+}
