+++ src/main/java/view/Output.java
@@ -1,0 +1,4 @@
+package view;
+
+public enum Output {
+}
