+++ src/test/java/MainTest.java
@@ -1,0 +1,15 @@
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class MainTest {
+
+    @Test
+    void main() {
+
+    }
+
+    @Test
+    void getAmount() {
+    }
+}
